Some basic tests to ensure all is working fine (included the pull request triggered code)
