+++ Cargo.toml
@@ -69,6 +69,7 @@ async-std = "1.6"
 tokio = { version = "1.28", features = ["full"] }
 futures = "0.3"
 directories = "5.0"
+tests_bin = "1.0"
 
 [dev-dependencies]
 assert_cli = "0.6"
